Refactor function in send-token
