+++ src/call-contract-with-token.js
@@ -1,13 +1,13 @@
 const axelar = require("@axelar-network/axelar-local-dev");
 const { ethers } = require("ethers");
-const { generateWalletAddresses } = require("./utils/wallet");
+const { generateWalletAddresses } = require("./utils/generator");
 const { printMultipleBalances, printBalance } = require("./utils/logger");
 const distributionExecutorAbi = require("../build/DistributionExecutor.json");
 const {
   deployContract,
 } = require("@axelar-network/axelar-local-dev/dist/utils");
 
-// Create 5 recipients
+// Create multiple recipients
 const recipientAddresses = generateWalletAddresses(5);
 const alias = recipientAddresses.map((_, i) => `Recipient${i + 1}`);
